chatbot/requirements.txt
pandas
python-dotenv
pypdf
unstructured
faiss-cpu
langchain==1.0.5
langchain-classic==1.0.0
langchain-text-splitters==1.0.0
langchain-community==0.4.1
langchain-openai==1.0.2
sentence-transformers==5.1.2
transformers==4.57.1
streamlit==1.51.0
botocore==1.34.69
boto3==1.34.69
supabase==2.24.0
